code review comments
In particular, move all the subscriber stuff into its own module and add
a big block comment to describe why it exists and what it does.
